% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connection.R
\name{print.netfs_connection}
\alias{print.netfs_connection}
\title{Print a netfs connection}
\usage{
\method{print}{netfs_connection}(x, ...)
}
\arguments{
\item{x}{A netfs connection.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a netfs connection
}
