% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dir.R
\name{dir_exists}
\alias{dir_exists}
\title{Test whether directories exist}
\usage{
dir_exists(path, con = NULL)
}
\arguments{
\item{path}{A local or remote path. For \code{dir_ls()}, a connection supplied as
the first argument is shorthand for listing that connection's root.}

\item{con}{\code{NULL} for local \code{fs} behavior, or a netfs connection.}
}
\value{
A logical vector.
}
\description{
Test whether directories exist
}
\seealso{
Other filesystem operations: 
\code{\link{dir_copy}()},
\code{\link{dir_create}()},
\code{\link{dir_delete}()},
\code{\link{dir_info}()},
\code{\link{dir_ls}()},
\code{\link{dir_map}()},
\code{\link{dir_tree}()},
\code{\link{dir_walk}()},
\code{\link{file_access}()},
\code{\link{file_chmod}()},
\code{\link{file_chown}()},
\code{\link{file_copy}()},
\code{\link{file_create}()},
\code{\link{file_delete}()},
\code{\link{file_download}()},
\code{\link{file_exists}()},
\code{\link{file_info}()},
\code{\link{file_move}()},
\code{\link{file_show}()},
\code{\link{file_touch}()},
\code{\link{file_transfer}()},
\code{\link{file_upload}()},
\code{\link{is_dir_empty}()},
\code{\link{is_file}()},
\code{\link{link_copy}()},
\code{\link{link_create}()},
\code{\link{link_delete}()},
\code{\link{link_path}()}
}
\concept{filesystem operations}
