% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{file_exists}
\alias{file_exists}
\title{Test whether files exist}
\usage{
file_exists(path, con = NULL)
}
\arguments{
\item{path}{A local or remote path.}

\item{con}{\code{NULL} for local \code{fs} behavior, or a netfs connection.}
}
\value{
A logical vector.
}
\description{
Test whether files exist
}
\seealso{
Other filesystem operations:
\code{\link[=dir_create]{dir_create()}},
\code{\link[=dir_delete]{dir_delete()}},
\code{\link[=dir_exists]{dir_exists()}},
\code{\link[=dir_ls]{dir_ls()}},
\code{\link[=file_copy]{file_copy()}},
\code{\link[=file_delete]{file_delete()}},
\code{\link[=file_download]{file_download()}},
\code{\link[=file_info]{file_info()}},
\code{\link[=file_move]{file_move()}},
\code{\link[=file_upload]{file_upload()}}
}
\concept{filesystem operations}
