% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/info.R
\name{file_info}
\alias{file_info}
\title{Retrieve file metadata}
\usage{
file_info(path, con = NULL, ...)
}
\arguments{
\item{path}{A local or remote path.}

\item{con}{\code{NULL} for local \code{fs} behavior, or a netfs connection.}

\item{...}{Arguments passed to \code{fs::dir_ls()} locally or to the backend.}
}
\value{
Locally, the result of \code{\link[fs:file_info]{fs::file_info()}}. Remotely, a tibble with
\code{path}, \code{type}, \code{size}, and \code{modification_time} columns.
}
\description{
Retrieve file metadata
}
\seealso{
Other filesystem operations:
\code{\link[=dir_create]{dir_create()}},
\code{\link[=dir_delete]{dir_delete()}},
\code{\link[=dir_exists]{dir_exists()}},
\code{\link[=dir_ls]{dir_ls()}},
\code{\link[=file_copy]{file_copy()}},
\code{\link[=file_delete]{file_delete()}},
\code{\link[=file_download]{file_download()}},
\code{\link[=file_exists]{file_exists()}},
\code{\link[=file_move]{file_move()}},
\code{\link[=file_upload]{file_upload()}}
}
\concept{filesystem operations}
